Split out DirTree store methods into LocalStore base class

+++ arcana/core/data/store/base.py
@@ -21,7 +21,8 @@ from arcana.core.utils.misc import (
     NestedContext,
 )
 from arcana.core.utils.packaging import list_subclasses
-from arcana.core.exceptions import ArcanaUsageError, ArcanaNameError
+from arcana.core.exceptions import ArcanaUsageError, ArcanaNameError, ArcanaError
+
 
 DS = ty.TypeVar("DS", bound="DataStore")
 
@@ -29,47 +30,11 @@ logger = logging.getLogger("arcana")
 
 
 if ty.TYPE_CHECKING:
-    from .space import DataSpace
-    from .set import DataTree
-    from .entry import DataEntry
-    from .row import DataRow
-
-
-@attrs.define(kw_only=True)
-class ExpDatatypeBlueprint:
-
-    datatype: type
-    filenames: list[str]
-
-
-@attrs.define(kw_only=True)
-class DerivBlueprint:
-
-    name: str
-    row_frequency: DataSpace
-    datatype: type
-    filenames: ty.List[str]
-
-
-@attrs.define(slots=False, kw_only=True)
-class TestDatasetBlueprint:
-
-    hierarchy: ty.List[DataSpace]
-    dim_lengths: ty.List[int]  # size of layers a-d respectively
-    files: ty.List[str]  # files present at bottom layer
-    id_inference: ty.List[ty.Tuple[DataSpace, str]] = attrs.field(
-        factory=list
-    )  # id_inference dict
-    expected_datatypes: ty.Dict[str, ExpDatatypeBlueprint] = attrs.field(
-        factory=dict
-    )  # expected formats
-    derivatives: ty.List[DerivBlueprint] = attrs.field(
-        factory=list
-    )  # files to insert as derivatives
-
-    @property
-    def space(self):
-        return type(self.hierarchy[0])
+    from ..set import Dataset
+    from ..tree import DataTree
+    from ..entry import DataEntry
+    from ..row import DataRow
+    from ..testing import TestDatasetBlueprint
 
 
 @attrs.define
@@ -84,7 +49,7 @@ class ConnectionManager(NestedContext):
     def enter(self):
         self.session = self.store.connect()
 
-    def disconnect(self):
+    def exit(self):
         self.store.disconnect(self.session)
         self.session = None
 
@@ -107,9 +72,11 @@ class DataStore(metaclass=ABCMeta):
 
     CONFIG_NAME = "stores"
     SUBPACKAGE = "data"
+    VERSION_KEY = "store-version"
+    VERSION = "1.0.0"
 
     @abstractmethod
-    def populate_tree(self, tree: DataTree):
+    def scan_tree(self, tree: DataTree):
         """
         Populates the nodes of the data tree with those found in the dataset
 
@@ -120,7 +87,7 @@ class DataStore(metaclass=ABCMeta):
         """
 
     @abstractmethod
-    def populate_row(self, row: DataRow):
+    def scan_row(self, row: DataRow):
         """
         Populate a row with all data entries found in the corresponding node in the data
         store (e.g. files within a directory, scans within an XNAT session).
@@ -132,14 +99,16 @@ class DataStore(metaclass=ABCMeta):
         """
 
     @abstractmethod
-    def get(self, entry: DataEntry) -> DataType:
+    def get(self, entry: DataEntry, datatype: type) -> DataType:
         """
         Gets the data item corresponding to the given entry
 
         Parameters
         ----------
-        entry: DataEntry
+        entry : DataEntry
             the data entry to update
+        datatype : type
+            the datatype to interpret the entry's item as
 
         Returns
         -------
@@ -252,12 +221,62 @@ class DataStore(metaclass=ABCMeta):
             A dct Dataset object that was saved in the data store
         """
 
+    def create_empty_dataset(
+        self,
+        id: str,
+        hierarchy: list[str],
+        row_ids: list[list[str]],
+        space: type,
+        name: str = None,
+        **kwargs,
+    ):
+        """create a new empty dataset within the store, used in test routines and
+        dataset imports
+
+        Parameters
+        ----------
+        id : str
+            ID for the newly created dataset
+        space : type(DataSpace)
+            the data space of the dataset
+        hierarchy : list[str]
+            the hierarchy of row frequencies the tree of the dataset
+        row_ids : list[list[str]]
+            the empty data rows to create within the newly created dataset. Outer list
+            corresponds to levels in the hierarchy, inner lists correspond to the row
+            ids repeated in each branch of the corresponding level of the data tree.
+        name : str, optional
+            name to give to the dataset definition returned
+        **kwargs
+            store sub-class relevant arguments
+        """
+        raise NotImplementedError(
+            f"{type(self).__name__} stores don't implement the `create_empty_dataset` "
+            "method"
+        )
+
+    def import_dataset(self, id: str, dataset: Dataset, **kwargs):
+        """Import a dataset from another store, transferring metadata and columns
+        defined on the original dataset
+
+        Parameters
+        ----------
+        id : str
+            the ID of the dataset within this store
+        dataset : Dataset
+            the dataset to import
+        **kwargs:
+            keyword arguments passed through to the `create_empty_dataset` method
+        """
+        raise NotImplementedError
+        # imported = self.create_empty_dataset(id, **kwargs)
+
     def connect(self):
         """
         If a connection session is required to the store manage it here
         """
 
-    def disconnect(self):
+    def disconnect(self, session):
         """
         If a connection session is required to the store manage it here
         """
@@ -357,9 +376,10 @@ class DataStore(metaclass=ABCMeta):
         del entries[name]
         cls.save_entries(entries)
 
-    def new_dataset(self, id, space=None, hierarchy=None, **kwargs):
+    def define_dataset(self, id, space=None, hierarchy=None, **kwargs):
         """
-        Returns a dataset from the XNAT repository
+        Creates a Arcana dataset definition for an existing data in the
+        data store.
 
         Parameters
         ----------
@@ -376,7 +396,7 @@ class DataStore(metaclass=ABCMeta):
         **kwargs:
             Keyword args passed on to the Dataset init method
         """
-        from .space import DataSpace
+        from ..space import DataSpace
 
         if not hierarchy:
             if space:
@@ -407,7 +427,44 @@ class DataStore(metaclass=ABCMeta):
         dataset = Dataset(id, store=self, space=space, hierarchy=hierarchy, **kwargs)
         return dataset
 
-    def load_dataset(self, id, name=None):
+    def save_dataset(self, dataset: Dataset, name: str = None):
+        """Save metadata in project definition file for future reference
+
+        Parameters
+        ----------
+        dataset : Dataset
+            the dataset to save
+        name : str, optional
+            the name for the definition to distinguish from other definitions on
+            the same data, by default None
+        """
+        definition = asdict(dataset, omit=["store", "name"])
+        definition[self.VERSION_KEY] = self.VERSION
+        if name is None:
+            name = dataset.name
+        self.save_dataset_definition(dataset.id, definition, name=name)
+
+    def load_dataset(self, id, name=None, **kwargs):
+        """Load an existing dataset definition
+
+        Parameters
+        ----------
+        id : str
+            ID of the dataset within the store
+        name : str, optional
+            name of the dataset definition, which distinguishes it from alternative
+            definitions on the same data, by default None
+
+        Returns
+        -------
+        Dataset
+            the loaded dataset
+
+        Raises
+        ------
+        KeyError
+            if the dataset is not found
+        """
         from arcana.core.data.set import (
             Dataset,
         )  # avoid circular imports it is imported here rather than at the top of the file
@@ -417,7 +474,9 @@ class DataStore(metaclass=ABCMeta):
         dct = self.load_dataset_definition(id, name)
         if dct is None:
             raise KeyError(f"Did not find a dataset '{id}@{name}'")
-        return fromdict(dct, id=id, name=name, store=self)
+        store_version = dct.pop(self.VERSION_KEY)
+        self.check_store_version(store_version)
+        return fromdict(dct, id=id, name=name, store=self, **kwargs)
 
     @classmethod
     def singletons(cls):
@@ -477,7 +536,14 @@ class DataStore(metaclass=ABCMeta):
         )
 
     @classmethod
-    def create_test_fsobject(cls, fname: str, dpath: Path, source_data: Path = None):
+    def create_test_fsobject(
+        cls,
+        fname: str,
+        dpath: Path,
+        source_data: Path = None,
+        source_fallback: bool = False,
+        escape_source_name: bool = True,
+    ):
         """For use in test routines, this classmethod creates a simple text file,
         zip file or nested directory at the given path
 
@@ -491,6 +557,11 @@ class DataStore(metaclass=ABCMeta):
         source_data : Path, optional
             path to a directory containing source data to use instead of the dummy
             data
+        source_fallback : bool
+            whether to fall back to the generated file if fname isn't in the source
+            data dir
+        escape_source_name : bool
+            whether to escape the source name or simple use the file name of the source
 
         Returns
         -------
@@ -501,39 +572,48 @@ class DataStore(metaclass=ABCMeta):
         dpath.mkdir(parents=True, exist_ok=True)
         if source_data is not None:
             src_path = source_data.joinpath(*fname.split("/"))
-            parts = fname.split(".")
-            out_path = dpath / (path2varname(parts[0]) + "." + ".".join(parts[1:]))
-            if src_path.is_dir():
-                shutil.copytree(src_path, out_path)
-            else:
-                shutil.copyfile(src_path, out_path, follow_symlinks=True)
+            if src_path.exists():
+                if escape_source_name:
+                    parts = fname.split(".")
+                    out_fname = path2varname(parts[0]) + "." + ".".join(parts[1:])
+                else:
+                    out_fname = Path(fname).name
+                out_path = dpath / out_fname
+                if src_path.is_dir():
+                    shutil.copytree(src_path, out_path)
+                else:
+                    shutil.copyfile(src_path, out_path, follow_symlinks=True)
+                return out_path
+            elif not source_fallback:
+                raise ArcanaError(
+                    f"Couldn't find {fname} in source data directory {source_data}"
+                )
+        out_path = dpath / fname
+        next_part = fname
+        if next_part.endswith(".zip"):
+            next_part = next_part.strip(".zip")
+        next_path = Path(next_part)
+        # Make double dir
+        if next_part.startswith("doubledir"):
+            (dpath / next_path).mkdir(exist_ok=True)
+            next_part = "dir"
+            next_path /= next_part
+        if next_part.startswith("dir"):
+            (dpath / next_path).mkdir(exist_ok=True)
+            next_part = "test.txt"
+            next_path /= next_part
+        if not next_path.suffix:
+            next_path = next_path.with_suffix(".txt")
+        if next_path.suffix == ".json":
+            contents = '{"a": 1.0}'
         else:
-            out_path = dpath / fname
-            next_part = fname
-            if next_part.endswith(".zip"):
-                next_part = next_part.strip(".zip")
-            next_path = Path(next_part)
-            # Make double dir
-            if next_part.startswith("doubledir"):
-                (dpath / next_path).mkdir(exist_ok=True)
-                next_part = "dir"
-                next_path /= next_part
-            if next_part.startswith("dir"):
-                (dpath / next_path).mkdir(exist_ok=True)
-                next_part = "test.txt"
-                next_path /= next_part
-            if not next_path.suffix:
-                next_path = next_path.with_suffix(".txt")
-            if next_path.suffix == ".json":
-                contents = '{"a": 1.0}'
-            else:
-                contents = fname
-            with open(dpath / next_path, "w") as f:
-                f.write(contents)
-            if fname.endswith(".zip"):
-                with zipfile.ZipFile(out_path, mode="w") as zfile, set_cwd(dpath):
-                    zfile.write(next_path)
-                (dpath / next_path).unlink()
+            contents = fname
+        with open(dpath / next_path, "w") as f:
+            f.write(contents)
+        if fname.endswith(".zip"):
+            with zipfile.ZipFile(out_path, mode="w") as zfile, set_cwd(dpath):
+                zfile.write(next_path)
+            (dpath / next_path).unlink()
         return out_path
 
     def make_test_dataset(
@@ -551,7 +631,7 @@ class DataStore(metaclass=ABCMeta):
         return self.access_test_dataset(blueprint, dataset_id)
 
     def access_test_dataset(self, blueprint, dataset_id):
-        dataset = self.new_dataset(
+        dataset = self.define_dataset(
             dataset_id,
             hierarchy=blueprint.hierarchy,
             id_inference=blueprint.id_inference,
@@ -568,3 +648,24 @@ class DataStore(metaclass=ABCMeta):
             for layer in blueprint.hierarchy:
                 ids[layer] = "".join(f"{b}{base_ids[b]}" for b in layer.span())
             yield ids
+
+    def check_store_version(self, store_version: str):
+        """Check whether version store used to save the dataset is compatible with the
+        current version of the software. Can be overridden by store subclasses where
+        appropriate
+
+        Parameters
+        ----------
+        store_version : str
+            version of the store used to save the dataset
+
+        Raises
+        ------
+        ArcanaError
+            if the saved version isn't compatible
+        """
+        if store_version != self.VERSION:
+            raise ArcanaError(
+                f"Stored version of dataset ({store_version}) does not match current "
+                f"version of {type(self).__name__} ({self.VERSION})"
+            )
